+++ src/pyproject2conda/__init__.py
@@ -2,15 +2,15 @@
 Top level API (:mod:`pyproject2conda`)
 ======================================
 """
+from importlib.metadata import PackageNotFoundError
+from importlib.metadata import version as _version
 
-from .parser import PyProject2Conda
-
-# updated versioning scheme
 try:
-    from ._version import __version__
-except Exception:  # pragma: no cover
+    __version__ = _version("pyproject2conda")
+except PackageNotFoundError:  # pragma: no cover
     __version__ = "999"
 
+from .parser import PyProject2Conda
 
 __author__ = """William P. Krekelberg"""
 __email__ = "wpk@nist.gov"