+++ tests/test_parser.py
@@ -189,12 +189,91 @@ dependencies:
     assert s == dedent(expected)
 
 
+def test_infer():
+    toml = dedent(
+        """\
+    [project]
+    name = "hello"
+    dependencies = [
+    "athing", # p2c: -p # a comment
+    "bthing", # p2c: -s bthing-conda
+    "cthing; python_version<'3.10'", # p2c: -c conda-forge
+    ]
+
+    [project.optional-dependencies]
+    test = [
+    "pandas",
+    "pytest", # p2c: -c conda-forge
+    ]
+    dev-extras = [
+    # p2c: -s additional-thing # this is an additional conda package
+    "matplotlib", # p2c: -s conda-matplotlib
+    ]
+    dev = [
+    "hello[test]",
+    "hello[dev-extras]",
+    ]
+    dist-pypi = [
+    "setuptools",
+    "build", # p2c: -p
+    ]
+
+
+    [tool.pyproject2conda]
+    channels = ['conda-forge']
+    """
+    )
+
+    d = parser.PyProject2Conda.from_string(toml)
+    with pytest.raises(ValueError):
+        d.to_conda_yaml(python_include="infer")
+
+
+def test_package_name():
+    toml = dedent(
+        """\
+    [project]
+    requires-python = ">=3.8,<3.11"
+    dependencies = [
+    "athing", # p2c: -p # a comment
+    "bthing", # p2c: -s bthing-conda
+    "cthing; python_version<'3.10'", # p2c: -c conda-forge
+    ]
+
+    [project.optional-dependencies]
+    test = [
+    "pandas",
+    "pytest", # p2c: -c conda-forge
+    ]
+    dev-extras = [
+    # p2c: -s additional-thing # this is an additional conda package
+    "matplotlib", # p2c: -s conda-matplotlib
+    ]
+    dev = [
+    "hello[test]",
+    "hello[dev-extras]",
+    ]
+    dist-pypi = [
+    "setuptools",
+    "build", # p2c: -p
+    ]
+
+
+    [tool.pyproject2conda]
+    channels = ['conda-forge']
+    """
+    )
+    d = parser.PyProject2Conda.from_string(toml)
+    with pytest.raises(ValueError):
+        d.to_conda_lists()
+
+
 def test_complete():
     toml = dedent(
         """\
     [project]
     name = "hello"
-    requires-python = ">=3.8,<3.11"
+    requires-python = ">=3.8, <3.11"
     dependencies = [
     "athing", # p2c: -p # a comment
     "bthing", # p2c: -s bthing-conda
@@ -262,6 +341,22 @@ dependencies:
       - athing
     """
     assert dedent(expected) == d.to_conda_yaml(python_include="infer")
+
+    # with remove spaces:
+    expected = """\
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.8, <3.11
+  - bthing-conda
+  - conda-forge::cthing
+  - pip
+  - pip:
+      - athing
+    """
+    assert dedent(expected) == d.to_conda_yaml(
+        python_include="infer", remove_whitespace=False
+    )
 
     expected = """\
 channels:
@@ -516,3 +611,70 @@ dependencies:
             f()
 
         assert f(allow_empty=True) == "No dependencies for this environment\n"
+
+
+def test_spaces():
+    toml = dedent(
+        """\
+    [project]
+    name = "hello"
+    requires-python = ">=3.8, <3.11"
+    dependencies = [
+    "athing >0.5", # p2c: -p # a comment
+    "bthing > 1.0", # p2c: -s 'bthing-conda > 2.0'
+    "cthing; python_version < '3.10'", # p2c: -c conda-forge
+    ]
+
+
+    [tool.pyproject2conda]
+    channels = ['conda-forge']
+    """
+    )
+
+    d = parser.PyProject2Conda.from_string(toml)
+
+    expected = """\
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.8, <3.11
+  - bthing-conda > 2.0
+  - conda-forge::cthing
+  - pip
+  - pip:
+      - athing >0.5
+    """
+    assert dedent(expected) == d.to_conda_yaml(
+        python_include="infer", remove_whitespace=False
+    )
+
+    expected = """\
+channels:
+  - conda-forge
+dependencies:
+  - python>=3.8,<3.11
+  - bthing-conda>2.0
+  - conda-forge::cthing
+  - pip
+  - pip:
+      - athing>0.5
+    """
+    assert dedent(expected) == d.to_conda_yaml(
+        python_include="infer", remove_whitespace=True
+    )
+
+    expected = """\
+    athing >0.5
+    bthing > 1.0
+    cthing; python_version < '3.10'
+    """
+
+    assert dedent(expected) == d.to_requirements(remove_whitespace=False)
+
+    expected = """\
+    athing>0.5
+    bthing>1.0
+    cthing;python_version<'3.10'
+    """
+
+    assert dedent(expected) == d.to_requirements(remove_whitespace=True)